Created 2D Camera View
Camera follows the player around the 2d world. When the player moves,
the camera and the mana/stamina bars stay/move with the camera (using
the distance offset from the player to the bar position, so offset is
constant to allow the bars to always be the same distance from the
player) in the Bar::arUpdate() function.

#Still to fix the target/lock system, mouse tracker